feat: email notification system

+++ api/src/services/appointmentService.js
@@ -1,8 +1,19 @@
 const appointmentRepo = require('../repositories/appointmentRepository');
+const { sendApprovalEmail } = require('./email-services/service-types/approve');
+const { sendDenialEmail } = require('./email-services/service-types/deny');
+const { sendCancellationEmail } = require('./email-services/service-types/cancel');
+const { sendAdminNotificationEmail } = require('./email-services/service-types/notify');
+const { sendConfirmationEmail } = require('./email-services/service-types/confirm');
 
 const createAppointment = async (appointmentData) => {
-    // (Check if date is available, send confirmation email to be added later)
-    return await appointmentRepo.create(appointmentData);
+    const newAppointment = await appointmentRepo.create(appointmentData);
+    
+    if (newAppointment) {
+        await sendAdminNotificationEmail(newAppointment);
+        await sendConfirmationEmail(newAppointment)
+    }
+    
+    return newAppointment;
 };
 
 const getAllAppointments = async () => {
@@ -18,19 +29,40 @@ const updateAppointment = async (id, updateData) => {
     return await appointmentRepo.update(id, updateData);
 };
 
-const denyAppointment = async (id) => {
-    // Refusal email could be added here later
-    return await appointmentRepo.deleteAppointment(id);
+const denyAppointment = async (id, message) => {
+    const appointment = await appointmentRepo.findById(id);
+    if (!appointment) return false;
+
+    const isDeleted = await appointmentRepo.deleteAppointment(id);
+
+    if (isDeleted) {
+        await sendDenialEmail(appointment, message);
+    }
+
+    return isDeleted;
 };
 
-const cancelAppointment = async (id) => {
-    // Cancellation hook could be added here later
-    return await appointmentRepo.deleteAppointment(id);
+const cancelAppointment = async (id, message) => {
+    const appointment = await appointmentRepo.findById(id);
+    if (!appointment) return false;
+
+    const isDeleted = await appointmentRepo.deleteAppointment(id);
+
+    if (isDeleted) {
+        await sendCancellationEmail(appointment, message);
+    }
+
+    return isDeleted;
 };
 
-const approveAppointment = async (id) => {
-    // Confirmation email to be added here later.
-    return await appointmentRepo.update(id, { approved: true });
+const approveAppointment = async (id, message) => {
+    const updatedAppointment = await appointmentRepo.update(id, { approved: true });
+    
+    if (updatedAppointment) {
+        await sendApprovalEmail(updatedAppointment, message);
+    }
+    
+    return updatedAppointment;
 };
 
 module.exports = {
